Extract updateVoteButtonStyles
