+++ src/services/voteService.js
@@ -59,3 +59,54 @@ export function getItemTypeNameFromId(typeId) {
   }
 }
 
+export function updateVoteButtonStyles(productId, voteType) {
+  const likeButtons = document.querySelectorAll(
+    `.like-button[data-product-id="${productId}"]`
+  );
+  const dislikeButtons = document.querySelectorAll(
+    `.dislike-button[data-product-id="${productId}"]`
+  );
+
+  if (likeButtons.length === 0 && dislikeButtons.length === 0) {
+    console.log(
+      `No vote buttons found anywhere on the page for product ID ${productId}.`
+    );
+    return;
+  }
+
+  let actionTaken = false;
+
+  likeButtons.forEach((button) => {
+    button.classList.remove("clicked");
+    if (voteType === "upvote") {
+      button.classList.add("clicked");
+      actionTaken = true;
+    }
+  });
+
+  dislikeButtons.forEach((button) => {
+    button.classList.remove("clicked");
+    if (voteType === "downvote") {
+      button.classList.add("clicked");
+      actionTaken = true;
+    }
+  });
+
+  if (actionTaken) {
+    if (voteType === "upvote") {
+      console.log(`Applied 'clicked' style to like button(s) for ${productId}`);
+    } else if (voteType === "downvote") {
+      console.log(
+        `Applied 'clicked' style to dislike button(s) for ${productId}`
+      );
+    }
+  } else if (
+    voteType === "none" &&
+    (likeButtons.length > 0 || dislikeButtons.length > 0)
+  ) {
+    console.log(
+      `Removed 'clicked' style from vote buttons for ${productId}`
+    );
+  }
+}
+
